+++ app/src/main/java/hu/bme/aut/fitary/ui/charts/ChartsFragment.kt
@@ -1,13 +1,12 @@
 package hu.bme.aut.fitary.ui.charts
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import hu.bme.aut.fitary.R
+import androidx.fragment.app.Fragment
 import hu.bme.aut.fitary.MainActivity
-import hu.bme.aut.fitary.adapter.ChartPagerAdapter
+import hu.bme.aut.fitary.R
 import kotlinx.android.synthetic.main.fragment_charts.view.*
 
 class ChartsFragment : Fragment() {
