+++ app/src/main/java/hu/bme/aut/fitary/ui/charts/barChart/BarChartViewModel.kt
@@ -1,4 +1,4 @@
-package hu.bme.aut.fitary.ui.barChart
+package hu.bme.aut.fitary.ui.charts.barChart
 
 import co.zsmb.rainbowcake.base.RainbowCakeViewModel
 import javax.inject.Inject
